# ZFS traverse_visitbp optimization to limit prefetch.