@adamdmoss 2021-03-29
I gave this a spin for a couple of day.  Nothing exploded.  Not sure I could measure a perf/mem gain/loss in casual use.

I have a _very mild_ reservation, which is that I would expect users to tune prefetch tuneables according to how their userspace workload behaves, i.e. how useful prefetch appears to be under reasonably predictable userspace test cases.  Whereas sanely tuning the traversal limit seems rather difficult, and depends on opaque zfs structural decisions.