@jsai20 2021-03-30
> I gave this a spin for a couple of day. Nothing exploded. Not sure I could measure a perf/mem gain/loss in casual use.
> 
> I have a _very mild_ reservation, which is that I would expect users to tune prefetch tuneables according to how their userspace workload behaves, i.e. how useful prefetch appears to be under reasonably predictable userspace test cases. Whereas sanely tuning the traversal limit seems rather difficult, and depends on opaque zfs structural decisions.

Thanks for the testing and your comment.
As per my understanding, On system with limited resources (IO bandwidth, System Memory) shared between block traversal (like zfs diff, zfs send) workload and primary user/application workload, its good to limit prefetch from traversal in one go to good enough number.
On High end systems, which has very low IO latency from underneath storage, has large amount of system memory (Larger size ARC) And Primary workload not consuming much of available IO band width and ARC, then probably its OK to do bulk Prefetch in one go, although not sure, if its really necessary to do bulk prefetch.
In Current code, Traversering an indirect block could do bulk prefetch upto 1024 blocks (As 128K size indirect block could contain 1024 block pointers). If needed to be do bulk prefetch for any user/application workload and high end system configuration, then current behavior of bulk prefetch can be restored by tunning "zfs_traverse_indirect_prefetch_limit" to value >=1024.