@amotin 2021-03-30
> I have a _very mild_ reservation, which is that I would expect users to tune prefetch tuneables according to how their userspace workload behaves, i.e. how useful prefetch appears to be under reasonably predictable userspace test cases. Whereas sanely tuning the traversal limit seems rather difficult, and depends on opaque zfs structural decisions.

@adamdmoss Unlike zfetch, this code is not about speculative prefetch, but more about read-ahead.  As I understand, most if not all blocks that are prefetched here should be read later, the only question is how much later.  With existing code it means 128MB for each indirect block level, which for some lets say L5 indirects of huge file means 1024 ^^ 5 data blocks ahead, that does not make much sense and can get evicted from ARC sooner that processing get to it.