@jsai20 2021-03-30
> I see sense in this change. Originally till [d7958b4](https://github.com/openzfs/zfs/commit/d7958b4cda27b82bd24dd4d4b2a539e26c62a409) ZFS used 8 times smaller indirect blocks (16KB instead of 128KB), that means 64 times less prefetch size here. We at iXsystems found this indirect block size pretty bad for random workloads, causing excessive read/write inflation, that is why we are still using 32KB now. But in general case I think this prefetch limit idea is good. Implementation does not look great to me though. Please see inline comments.

Thanks for the review. Please see responses to comments inline.