@jsai20 2021-04-06
> > I have no more objections. It would still be good to see some benchmarks, showing that the reduced prefetch depth and removed prefetch of the first block do not cause performance regressions.
> 
> Thanks @amotin for the review.
> If you are Ok, We can start prefetch from index=1 (remove 1st block prefetch) just for prefetchlimit = 1 and for rest, can start at index = 0 only; So, that any possibility of regression by removing 1st block prefetch is avoided.
> 
> ```
>     prefetchlimit = zfs_traverse_indirect_prefetch_limit;
>     if (prefetchlimit == 1)
>             pidx = 1;
>     else
>             pidx = 0;
> ```
> 
> And, with this, for setting next value of "ptidx", inside "for loop", would now need to do, ptidx = MAX(pidx, i + 1);

Just update this change.