@jsai20 2021-04-06
> > Just updated this change.
> 
> I don't like this update. It complicates the code without proven motivation. Lets return back and run at least some reasonable benchmarks, comparing behavior with this patch and without.

Ok.