@jsai20 2021-04-07
> > > Just updated this change.
> > 
> > 
> > I don't like this update. It complicates the code without proven motivation. Lets return back and run at least some reasonable benchmarks, comparing behavior with this patch and without.
> 
> Ok.

**Perf Testing:**
This optimisation (Limiting prefetch done during traverse of indirect blocks) is primarily meant to minimize any negative impact to other primary workloads running on system in parallel to workloads doing block traversal (zfs send). Negative impacts possible due to bulk prefetch done by traversal code while traversing indirect blocks.
And considering necessary/limited prefetch is still done by traverse code, it shouldn’t have any negative impact (No regression) to zfs send perf. 

Ran couple of basic test scenarios to validate no negative impact (No regression) to zfs send with this optimization.

**Test System Memory and CPU Config:**

Total online memory:      12G
CPU(s):                4
Model name:            Intel(R) Xeon(R) CPU E5-2620 0 @ 2.00GHz


**Test Scenario 1:**
Source Dataset containing Small files with number of indirect blocks like,  L2 = 1,  L1 = 2.
Without optimization, two L1 blocks  (at index 0 and index 1)  gets prefetched during traversal, while with optimization only one block (at Index 1, skipping block at index 0) gets prefetched.
Taken 2 runs both with and without optimization to account for variation. 
With averaging out numbers from 2 runs, Performance looks similar, with and without optimization.


$  sudo dd if=/dev/urandom of=/testpool/testds/134M-file bs=1k count=131200
sudo rm 131200+0 records in
131200+0 records out
134348800 bytes (134 MB) copied, 7.55323 s, 17.8 MB/s
$ ls -i /testpool/testds/134M-file 
256 /testpool/testds/134M-file
$

$ sudo zdb -ddddd testpool/testds 256 |egrep " L0 " |wc -l
1025
$ sudo zdb -ddddd testpool/testds 256 |egrep " L1 " |wc -l
2
$ sudo zdb -ddddd testpool/testds 256 |egrep " L2 " |wc -l
1
$

$ for i in `seq 1 200`;do sudo dd if=/dev/urandom of=/testpool/testds/134M-file-$i bs=1k count=131200;done

$ ls /testpool/testds/
134M-file      134M-file-119  134M-file-14   134M-file-160  134M-file-181  134M-file-21  134M-file-42  134M-file-63  134M-file-84
134M-file-1    134M-file-12   134M-file-140  134M-file-161  134M-file-182  134M-file-22  134M-file-43  134M-file-64  134M-file-85
134M-file-10   134M-file-120  134M-file-141  134M-file-162  134M-file-183  134M-file-23  134M-file-44  134M-file-65  134M-file-86
134M-file-100  134M-file-121  134M-file-142  134M-file-163  134M-file-184  134M-file-24  134M-file-45  134M-file-66  134M-file-87
134M-file-101  134M-file-122  134M-file-143  134M-file-164  134M-file-185  134M-file-25  134M-file-46  134M-file-67  134M-file-88
134M-file-102  134M-file-123  134M-file-144  134M-file-165  134M-file-186  134M-file-26  134M-file-47  134M-file-68  134M-file-89
134M-file-103  134M-file-124  134M-file-145  134M-file-166  134M-file-187  134M-file-27  134M-file-48  134M-file-69  134M-file-9
134M-file-104  134M-file-125  134M-file-146  134M-file-167  134M-file-188  134M-file-28  134M-file-49  134M-file-7   134M-file-90
134M-file-105  134M-file-126  134M-file-147  134M-file-168  134M-file-189  134M-file-29  134M-file-5   134M-file-70  134M-file-91
134M-file-106  134M-file-127  134M-file-148  134M-file-169  134M-file-19   134M-file-3   134M-file-50  134M-file-71  134M-file-92
134M-file-107  134M-file-128  134M-file-149  134M-file-17   134M-file-190  134M-file-30  134M-file-51  134M-file-72  134M-file-93
134M-file-108  134M-file-129  134M-file-15   134M-file-170  134M-file-191  134M-file-31  134M-file-52  134M-file-73  134M-file-94
134M-file-109  134M-file-13   134M-file-150  134M-file-171  134M-file-192  134M-file-32  134M-file-53  134M-file-74  134M-file-95
134M-file-11   134M-file-130  134M-file-151  134M-file-172  134M-file-193  134M-file-33  134M-file-54  134M-file-75  134M-file-96
134M-file-110  134M-file-131  134M-file-152  134M-file-173  134M-file-194  134M-file-34  134M-file-55  134M-file-76  134M-file-97
134M-file-111  134M-file-132  134M-file-153  134M-file-174  134M-file-195  134M-file-35  134M-file-56  134M-file-77  134M-file-98
134M-file-112  134M-file-133  134M-file-154  134M-file-175  134M-file-196  134M-file-36  134M-file-57  134M-file-78  134M-file-99
134M-file-113  134M-file-134  134M-file-155  134M-file-176  134M-file-197  134M-file-37  134M-file-58  134M-file-79
134M-file-114  134M-file-135  134M-file-156  134M-file-177  134M-file-198  134M-file-38  134M-file-59  134M-file-8
134M-file-115  134M-file-136  134M-file-157  134M-file-178  134M-file-199  134M-file-39  134M-file-6   134M-file-80
134M-file-116  134M-file-137  134M-file-158  134M-file-179  134M-file-2    134M-file-4   134M-file-60  134M-file-81
134M-file-117  134M-file-138  134M-file-159  134M-file-18   134M-file-20   134M-file-40  134M-file-61  134M-file-82
134M-file-118  134M-file-139  134M-file-16   134M-file-180  134M-file-200  134M-file-41  134M-file-62  134M-file-83
$
$ sudo zfs snapshot testpool/testds@snap1


**Without Optimization:**
$ cat /sys/module/zfs/parameters/zfs_traverse_indirect_prefetch_limit 
cat: /sys/module/zfs/parameters/zfs_traverse_indirect_prefetch_limit: No such file or directory


**1st Run:**
$ time sudo zfs send -v -c testpool/testds@snap1 >/dev/null 
full send of testpool/testds@snap1 estimated size is 25.2G
total estimated size is 25.2G
TIME        SENT   SNAPSHOT testpool/testds@snap1
06:12:14    149M   testpool/testds@snap1
06:12:15    253M   testpool/testds@snap1
06:12:16    393M   testpool/testds@snap1
06:12:17    523M   testpool/testds@snap1
06:12:18    695M   testpool/testds@snap1
….
06:12:34   1.80G   testpool/testds@snap1
06:12:35   1.86G   testpool/testds@snap1
06:12:36   1.96G   testpool/testds@snap1
….
06:16:54   25.0G   testpool/testds@snap1
06:16:55   25.0G   testpool/testds@snap1
06:16:56   25.1G   testpool/testds@snap1

real    4m44.335s
user    0m0.170s
sys     0m8.695s
-------------------------------------------------------
**2nd run (After Reboot):**
$ time sudo zfs send -v -c testpool/testds@snap1 >/dev/null 
full send of testpool/testds@snap1 estimated size is 25.2G
total estimated size is 25.2G
TIME        SENT   SNAPSHOT testpool/testds@snap1
07:05:27    115M   testpool/testds@snap1
07:05:28    244M   testpool/testds@snap1
07:05:29    341M   testpool/testds@snap1
07:05:30    416M   testpool/testds@snap1
07:05:31    515M   testpool/testds@snap1
….
07:05:50   1.68G   testpool/testds@snap1
07:05:51   1.77G   testpool/testds@snap1
07:05:52   1.82G   testpool/testds@snap1
….
07:10:39   25.0G   testpool/testds@snap1
07:10:40   25.1G   testpool/testds@snap1
07:10:41   25.1G   testpool/testds@snap1

real    5m16.421s
user    0m0.174s
sys     0m8.571s
-------------------------------------------------------

**With Optimization:**
$ cat /sys/module/zfs/parameters/zfs_traverse_indirect_prefetch_limit 
32

**1st run:**
$ time sudo zfs send -v -c testpool/testds@snap1 >/dev/null 
full send of testpool/testds@snap1 estimated size is 25.2G
total estimated size is 25.2G
TIME        SENT   SNAPSHOT testpool/testds@snap1
06:34:34    169M   testpool/testds@snap1
06:34:35    290M   testpool/testds@snap1
06:34:36    384M   testpool/testds@snap1
06:34:37    534M   testpool/testds@snap1
06:34:38    686M   testpool/testds@snap1
….
06:34:54   1.78G   testpool/testds@snap1
06:34:55   1.84G   testpool/testds@snap1
06:34:56   1.90G   testpool/testds@snap1
….
06:39:15   25.0G   testpool/testds@snap1
06:39:16   25.1G   testpool/testds@snap1
06:39:17   25.2G   testpool/testds@snap1

real    4m45.321s
user    0m0.165s
sys     0m8.708s
-------------------------------------------------------

**2nd run (After Reboot):**
$ time sudo zfs send -v -c testpool/testds@snap1 >/dev/null 
full send of testpool/testds@snap1 estimated size is 25.2G
total estimated size is 25.2G
TIME        SENT   SNAPSHOT testpool/testds@snap1
06:52:45    160M   testpool/testds@snap1
06:52:46    320M   testpool/testds@snap1
06:52:47    390M   testpool/testds@snap1
06:52:48    522M   testpool/testds@snap1
06:52:49    686M   testpool/testds@snap1
…..
06:53:06   1.78G   testpool/testds@snap1
06:53:07   1.83G   testpool/testds@snap1
06:53:08   1.88G   testpool/testds@snap1
…..
06:57:51   25.0G   testpool/testds@snap1
06:57:52   25.0G   testpool/testds@snap1
06:57:53   25.1G   testpool/testds@snap1

real    5m9.334s
user    0m0.157s
sys     0m8.635s
------------------------------------------------------------------


**Test Scenario 2:**
Test Dataset containing a large file such that it contains multiple number of indirect blocks at L1 pointed by L2.
 It tests Bulk prefetch done without optimization, while limited prefetch is done with optimization.

Taken 2 runs both with and without optimization, to account variation between runs.
Performance Number are stable across runs and its same both with and without optimization.

$ sudo dd if=/dev/urandom of=/testpool/testds/large-file bs=1k count=26240000
26240000+0 records in
26240000+0 records out
26869760000 bytes (27 GB) copied, 1483.99 s, 18.1 MB/s
$
$ ls /testpool/testds/
large-file
$ ls -i /testpool/testds/large-file
2 /testpool/testds/large-file
$ sudo zdb -ddddd testpool/testds 2 |egrep " L0 " |wc -l
205000
$ sudo zdb -ddddd testpool/testds 2 |egrep " L1 " |wc -l
201
$ sudo zdb -ddddd testpool/testds 2 |egrep " L2 " |wc -l
1
$


**Without Optimization:**
$ cat /sys/module/zfs/parameters/zfs_traverse_indirect_prefetch_limit 
cat: /sys/module/zfs/parameters/zfs_traverse_indirect_prefetch_limit: No such file or directory

**1st Run:**
$ time sudo zfs send -v -c testpool/testds@snap1 >/dev/null 
full send of testpool/testds@snap1 estimated size is 25.1G
total estimated size is 25.1G
TIME        SENT   SNAPSHOT testpool/testds@snap1
07:54:52   82.9M   testpool/testds@snap1
07:54:53    195M   testpool/testds@snap1
07:54:54    269M   testpool/testds@snap1
07:54:55    342M   testpool/testds@snap1
07:54:56    420M   testpool/testds@snap1
…..
07:55:13   1.72G   testpool/testds@snap1
07:55:14   1.87G   testpool/testds@snap1
07:55:15   1.99G   testpool/testds@snap1
….
07:59:05   24.9G   testpool/testds@snap1
07:59:06   24.9G   testpool/testds@snap1
07:59:07   25.0G   testpool/testds@snap1

real    4m17.322s
user    0m0.166s
sys     0m8.632s
-------------------------------------------------------

**2nd Run (After reboot):**
$ time sudo zfs send -v -c testpool/testds@snap1 >/dev/null                                                
full send of testpool/testds@snap1 estimated size is 25.1G                                                                                   
total estimated size is 25.1G                                                                                                                
TIME        SENT   SNAPSHOT testpool/testds@snap1                                                                                            
08:06:03   79.6M   testpool/testds@snap1                                                                                                     
08:06:04    189M   testpool/testds@snap1                                                                                                     
08:06:05    266M   testpool/testds@snap1                                                                                                     
08:06:06    334M   testpool/testds@snap1                                                                                                     
08:06:07    408M   testpool/testds@snap1   
…..
08:06:23   1.67G   testpool/testds@snap1                                                                                                     
08:06:24   1.84G   testpool/testds@snap1                                                                                                     
08:06:25   1.98G   testpool/testds@snap1
…..
08:10:20   24.9G   testpool/testds@snap1
08:10:21   25.0G   testpool/testds@snap1
08:10:22   25.1G   testpool/testds@snap1

real    4m21.343s
user    0m0.187s
sys     0m8.701s
-------------------------------------------------------

**With Optimization:**
$ cat /sys/module/zfs/parameters/zfs_traverse_indirect_prefetch_limit 
32

**1st Run:**
$ time sudo zfs send -v -c testpool/testds@snap1 >/dev/null
full send of testpool/testds@snap1 estimated size is 25.1G                                                                                   
total estimated size is 25.1G                                                                                                                
TIME        SENT   SNAPSHOT testpool/testds@snap1                                                                                            
08:18:13   85.8M   testpool/testds@snap1                                                                                                     
08:18:14    199M   testpool/testds@snap1                                                                                                     
08:18:15    278M   testpool/testds@snap1                                                                                                     
08:18:16    353M   testpool/testds@snap1                                                                                                     
08:18:17    426M   testpool/testds@snap1                                                                                                     
…..
08:18:30   1.34G   testpool/testds@snap1                                                                                                     
08:18:31   1.45G   testpool/testds@snap1                                                                                                     
08:18:32   1.62G   testpool/testds@snap1                     
…..
08:22:30   24.9G   testpool/testds@snap1
08:22:31   25.0G   testpool/testds@snap1
08:22:32   25.0G   testpool/testds@snap1

real    4m21.322s
user    0m0.162s
sys     0m8.604s
-------------------------------------------------------

**2nd Run (After Reboot):**
$ time sudo zfs send -v -c testpool/testds@snap1 >/dev/null
full send of testpool/testds@snap1 estimated size is 25.1G
total estimated size is 25.1G
TIME        SENT   SNAPSHOT testpool/testds@snap1
08:38:08   77.3M   testpool/testds@snap1
08:38:09    169M   testpool/testds@snap1
08:38:10    252M   testpool/testds@snap1
08:38:11    334M   testpool/testds@snap1
08:38:12    405M   testpool/testds@snap1
…..
08:38:28   1.66G   testpool/testds@snap1
08:38:29   1.83G   testpool/testds@snap1
08:38:30   1.97G   testpool/testds@snap1
…..
08:42:26   24.9G   testpool/testds@snap1
08:42:27   25.0G   testpool/testds@snap1
08:42:28   25.1G   testpool/testds@snap1

real    4m21.309s
user    0m0.157s
sys     0m8.525s
-------------------------------------------------------